requirements.txt
pyyaml==6.0.1
pytest==7.4.3